Extend rotor arms to add guards.
The guards surround the outer edges of the rotors and should
prevent them from getting damaged if the drone bumps into a wall.

Also adjust the rotor arms to be exactly equal lengths, as they
were previously not exactly equal. Turn off relative SVG coordinates
so that the design should load correctly in CorelDraw.